third_party/blink/web_tests/external/wpt/encoding/textdecoder-eof.any-expected.txt

This is a testharness.js-based test.
[FAIL] TextDecoder end-of-queue handling
  assert_equals: expected "\\ufffd" but got "\\ufffd\\ufffd"
[FAIL] TextDecoder end-of-queue handling using stream: true
  assert_equals: expected "" but got "\\ufffd"
Harness: the test ran to completion.
